chore(Rockerbox): move sdk loader to different file
